FlaskApp/requirements.txt
flask
pickle-mixin
Pillow
